+++ runtime/ms-rest-azure/lib/credentials/applicationTokenCredentials.js
@@ -28,7 +28,14 @@ const AzureEnvironment = require('../azureEnvironment');
     /* jshint validthis: true */
     let self = this;
     let resource = self.environment.activeDirectoryResourceId;
-    if (self.tokenAudience && self.tokenAudience.toLowerCase() === 'graph') resource = self.environment.activeDirectoryGraphResourceId;
+    if (self.tokenAudience)  {
+      resource = self.tokenAudience;
+      if (self.tokenAudience.toLowerCase() === 'graph') {
+        resource = self.environment.activeDirectoryGraphResourceId;
+      } else if (self.tokenAudience.toLowerCase() === 'batch') {
+        resource = self.environment.batchResourceId;
+      }
+    }
     self.context.acquireToken(resource, null, self.clientId, function (err, result) {
       if (err) {
         //make sure to remove the stale token from the tokencache. ADAL gives the same error message "Entry not found in cache."
@@ -56,8 +63,8 @@ const AzureEnvironment = require('../azureEnvironment');
  * @param {string} domain The domain or tenant id containing this application.
  * @param {string} secret The authentication secret for the application.
  * @param {object} [options] Object representing optional parameters.
- * @param {string} [options.tokenAudience] The audience for which the token is requested. Valid value is 'graph'. If tokenAudience is provided 
- * then domain should also be provided its value should not be the default 'common' tenant. It must be a string (preferrably in a guid format).
+ * @param {string} [options.tokenAudience] The audience for which the token is requested. Valid values are 'graph', 'batch' or any other resource like 'https://vault.azure.com/'.
+ * If tokenAudience is 'graph' then domain should also be provided and its value should not be the default 'common' tenant. It must be a string (preferrably in a guid format).
  * @param {AzureEnvironment} [options.environment] The azure environment to authenticate with.
  * @param {string} [options.authorizationScheme] The authorization scheme. Default value is 'bearer'.
  * @param {object} [options.tokenCache] The token cache. Default value is the MemoryCache object from adal.
@@ -92,14 +99,9 @@ class ApplicationTokenCredentials {
       options.tokenCache = new adal.MemoryCache();
     }
 
-    if (options.tokenAudience) {
-      if (options.tokenAudience.toLowerCase() !== 'graph') {
-        throw new Error('Valid value for \'tokenAudience\' is \'graph\'.');
-      }
-      if (domain.toLowerCase() === 'common') {
-        throw new Error('If the tokenAudience is specified as \'graph\' then \'domain\' cannot be the default \'commmon\' tenant. ' +
-          'It must be the actual tenant (preferrably a string in a guid format).');
-      }
+    if (options.tokenAudience && options.tokenAudience.toLowerCase() === 'graph' && domain.toLowerCase() === 'common') {
+      throw new Error('If the tokenAudience is specified as \'graph\' then \'domain\' cannot be the default \'commmon\' tenant. ' +
+        'It must be the actual tenant (preferrably a string in a guid format).');
     }
 
     this.tokenAudience = options.tokenAudience;
@@ -131,7 +133,14 @@ class ApplicationTokenCredentials {
         } else {
           //Some error occured in retrieving the token from cache. May be the cache was empty or the access token expired. Let's try again.
           let resource = self.environment.activeDirectoryResourceId;
-          if (self.tokenAudience && self.tokenAudience.toLowerCase() === 'graph') resource = self.environment.activeDirectoryGraphResourceId;
+          if (self.tokenAudience)  {
+            resource = self.tokenAudience;
+            if (self.tokenAudience.toLowerCase() === 'graph') {
+              resource = self.environment.activeDirectoryGraphResourceId;
+            } else if (self.tokenAudience.toLowerCase() === 'batch') {
+              resource = self.environment.batchResourceId;
+            }
+          }
           self.context.acquireTokenWithClientCredentials(resource, self.clientId, self.secret, function (err, tokenResponse) {
             if (err) {
               return callback(new Error('Failed to acquire token for application with the provided secret. \n' + err));