+++ runtime/ms-rest-azure/lib/credentials/deviceTokenCredentials.js
@@ -21,8 +21,8 @@ const AzureEnvironment = require('../azureEnvironment');
  * @param {string} [options.username] The user name for account in the form: 'user@example.com'.
  * @param {AzureEnvironment} [options.environment] The azure environment to authenticate with. Default environment is "Azure" popularly known as "Public Azure Cloud".
  * @param {string} [options.domain] The domain or tenant id containing this application. Default value is 'common'
- * @param {string} [options.tokenAudience] The audience for which the token is requested. Valid value is 'graph'. If tokenAudience is provided 
- * then domain should also be provided and its value should not be the default 'common' tenant. It must be a string (preferrably in a guid format). 
+ * @param {string} [options.tokenAudience] The audience for which the token is requested. Valid values are 'graph', 'batch' or any other resource like 'https://vault.azure.com/'.
+ * If tokenAudience is 'graph' then domain should also be provided and its value should not be the default 'common' tenant. It must be a string (preferrably in a guid format). 
  * @param {string} [options.clientId] The active directory application client id. 
  * See {@link https://azure.microsoft.com/en-us/documentation/articles/active-directory-devquickstarts-dotnet/ Active Directory Quickstart for .Net} 
  * for an example.
@@ -59,14 +59,9 @@ class DeviceTokenCredentials {
       options.tokenCache = new adal.MemoryCache();
     }
 
-    if (options.tokenAudience) {
-      if (options.tokenAudience.toLowerCase() !== 'graph') {
-        throw new Error('Valid value for \'tokenAudience\' is \'graph\'.');
-      }
-      if (options.domain.toLowerCase() === 'common') {
-        throw new Error('If the tokenAudience is specified as \'graph\' then \'domain\' cannot be the default \'commmon\' tenant. ' +
-          'It must be the actual tenant (preferrably a string in a guid format).');
-      }
+    if (options.tokenAudience && options.tokenAudience.toLowerCase() === 'graph' && options.domain.toLowerCase() === 'common') {
+      throw new Error('If the tokenAudience is specified as \'graph\' then \'domain\' cannot be the default \'commmon\' tenant. ' +
+        'It must be the actual tenant (preferrably a string in a guid format).');
     }
 
     this.tokenAudience = options.tokenAudience;
@@ -76,14 +71,21 @@ class DeviceTokenCredentials {
     this.clientId = options.clientId;
     this.authorizationScheme = options.authorizationScheme;
     this.tokenCache = options.tokenCache;
-    var authorityUrl = this.environment.activeDirectoryEndpointUrl + this.domain;
+    let authorityUrl = this.environment.activeDirectoryEndpointUrl + this.domain;
     this.context = new adal.AuthenticationContext(authorityUrl, this.environment.validateAuthority, this.tokenCache);
   }
 
   retrieveTokenFromCache(callback) {
-    var self = this;
-    var resource = self.environment.activeDirectoryResourceId;
-    if (self.tokenAudience && self.tokenAudience.toLowerCase() === 'graph') resource = self.environment.activeDirectoryGraphResourceId;
+    let self = this;
+    let resource = self.environment.activeDirectoryResourceId;
+    if (self.tokenAudience)  {
+      resource = self.tokenAudience;
+      if (self.tokenAudience.toLowerCase() === 'graph') {
+        resource = self.environment.activeDirectoryGraphResourceId;
+      } else if (self.tokenAudience.toLowerCase() === 'batch') {
+        resource = self.environment.batchResourceId;
+      }
+    }
     self.context.acquireToken(resource, self.username, self.clientId, function (err, result) {
       if (err) return callback(err);
       return callback(null, result.tokenType, result.accessToken);