+++ runtime/ms-rest-azure/Changelog.md
@@ -1,3 +1,7 @@
+### 2.5.3 (01/31/2018)
+- Opened up the optional parameter `tokenAudience` for several login methods to accept `'graph'`, `'batch'` or any other resource uri like `'https://vault.azure.com'`.
+- Added `batchResourceId` in all the supported AzureEnvironments.
+
 ### 2.5.2 (01/31/2018)
 - Migrated `KeyVaultCredentials` from keyvault data-plane and added support for various auth consumption.